requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.31
passlib[bcrypt]==1.7.4
itsdangerous==2.2.0
python-dotenv==1.0.1
openai==1.37.0
pydantic==1.10.15
pytest==8.3.2
httpx==0.27.0
black==24.8.0
ruff==0.5.5

